правки по дз
