+++ hw05/src/main/java/ru/otus/hw05/repository/JdbcGenreRepository.java
@@ -8,7 +8,7 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcOperations;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import ru.otus.hw05.model.Author;
+import ru.otus.hw05.model.Genre;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -19,39 +19,40 @@ import java.util.Optional;
 
 @Repository
 @RequiredArgsConstructor
-public class AuthorRepositoryJdbc implements AuthorRepository {
+public class JdbcGenreRepository implements GenreRepository {
 
-    private static class AuthorMapper implements RowMapper<Author> {
+    private static class GenreMapper implements RowMapper<Genre> {
+
         @Override
-        public Author mapRow(ResultSet resultSet, int rowNumber) throws SQLException {
+        public Genre mapRow(ResultSet resultSet, int rowNumber) throws SQLException {
             long id = resultSet.getLong("id");
             String name = resultSet.getString("name");
-            return new Author(id, name);
+            return new Genre(id, name);
         }
+
     }
 
     private final NamedParameterJdbcOperations namedParameterJdbcOperations;
 
-
     @Override
-    public List<Author> findAll() {
+    public List<Genre> findAll() {
         return namedParameterJdbcOperations.getJdbcOperations()
-                .query("select id, name from authors", new AuthorMapper());
+                .query("select g.id, g.name from genres g", new GenreMapper());
     }
 
     @Override
-    public Optional<Author> findById(long id) {
+    public Optional<Genre> findById(long id) {
         try {
             return Optional.ofNullable(namedParameterJdbcOperations.queryForObject(
-                    "select id, name from authors where id = :id",
-                    Map.of("id", id), new AuthorMapper()));
+                    "select id, name from genres where id=:id",
+                    Map.of("id", id), new GenreMapper()));
         } catch (DataAccessException e) {
             return Optional.empty();
         }
     }
 
     @Override
-    public Author save(Author genre) {
+    public Genre save(Genre genre) {
         if (genre.getId() == null) {
             return insert(genre);
         }
@@ -61,23 +62,23 @@ public class AuthorRepositoryJdbc implements AuthorRepository {
     @Override
     public long deleteById(long id) {
         return namedParameterJdbcOperations.update(
-                "delete from authors where id = :id", Map.of("id", id));
+                "delete from genres where id = :id", Map.of("id", id));
     }
 
-    private Author insert(Author author) {
+    private Genre insert(Genre genre) {
         MapSqlParameterSource parameterSource = new MapSqlParameterSource();
-        parameterSource.addValue("name", author.getName());
+        parameterSource.addValue("name", genre.getName());
         KeyHolder keyHolder = new GeneratedKeyHolder();
         namedParameterJdbcOperations.update(
-                "insert into authors (name) values (:name)", parameterSource, keyHolder, new String[]{"id"});
-        author.setId(Objects.requireNonNull(keyHolder.getKey()).longValue());
-        return author;
+                "insert into genres (name) values (:name)", parameterSource, keyHolder, new String[]{"id"});
+        genre.setId(Objects.requireNonNull(keyHolder.getKey()).longValue());
+        return genre;
     }
 
-    private Author update(Author author) {
+    private Genre update(Genre genre) {
         namedParameterJdbcOperations.update(
-                "update authors g set g.name = :name where g.id = :id",
-                Map.of("id", author.getId(), "name", author.getName()));
-        return author;
+                "update genres g set g.name = :name where g.id = :id",
+                Map.of("id", genre.getId(), "name", genre.getName()));
+        return genre;
     }
 }
